requirements.txt
langchain==0.0.232
tiktoken==0.4.0
pywikibot==8.2.0
mwparserfromhell==0.6.4
pandas==2.0.3
hydra-core==1.3.2
python-dotenv==1.0.0
pypdf==3.12.0
tqdm==4.65.0
